`Added method to assign results in Classic discipline`
` `

`A new method `AssignResultsBasedonClassicRuns` has been added to the `Classic.cs` file in the `FreestyleSlalomCompetitionManager.BL.Models.Disciplines.Classic` namespace. This method assigns results based on the competitors' final marks in the first round of the competition. It iterates over the runs in the first round, ordered by the final mark of each competitor, and adds each competitor to the results list, incrementing the order each time. No changes were made to the `ClassicTests.cs` file; '+' symbols at the end of the file were identified as a formatting error.`