Moved `Battle` class and added `BattleGroup` and `BattleRound` classes
Moved the `Battle` class to a new namespace `FreestyleSlalomCompetitionManager.BL.Models.Disciplines.Battle`. Added a new class `BattleGroup` with a `Competitors` property, which is an empty dictionary with `Competitor` as the key and `int` as the value. Also added a new class `BattleRound` that inherits from `BaseRound` and has a `Runs` property, which is an empty list of `BattleGroup`.