Updated Battle logic and tests, modified Competitor creation
In `Battle.cs`, the `InitializeBattle` method was updated to pass `Competitors` as a parameter to the `BattleRound` constructor. A new constructor was added in `BattleRound.cs` that accepts a `SortedDictionary<int, Competitor>` and calls the `CreateGroups` and `AssignCompetitorsToGroups` methods. The `AssignCompetitorsToGroups` method was refactored to use a `foreach` loop.

In `BaseTest.cs`, the `CreateCompetitor` and `CreateListOfCompetitors` methods were updated to take `AgeCategory?` and `SexCategory?` as parameters. Calls to these methods in `FreestyleSlalomCompetitionManagerRunnerTests.cs` and `CompetitionTests.cs` were updated accordingly.

Two new test methods were added in `BattleTests.cs` to verify the functionality of the `AssignCompetitors` and `InitializeBattle` methods.