`Added ExportRoundStartList class for Excel export`
`Added a new class `ExportRoundStartList` in the `FreestyleSlalomCompetitionManager.BL.Exports` namespace. This class is responsible for exporting the start list of a battle round in a freestyle slalom competition to an Excel file using the `OfficeOpenXml` library. The `ExportBattleRoundStartList` method creates a new Excel worksheet titled "Skater registration data", iterates over the groups in the round, and writes the group number, start list, and competitor details to the worksheet. The method also applies a thin border to the cells containing the competitor details. However, the method does not save the created Excel package to a file or stream, which might be an oversight or the code might be incomplete.`