Subject: Added export functionality and updated tests
This commit includes several updates to the freestyle slalom competition management system.

In `FreestyleSlalomCompetitionManagerRunner.cs`, a new case "exportstartinglists" has been added to the switch statement, which calls the new `ExportStartingLists()` method. This method checks if the current competition exists and if there are any disciplines and competitors, and if so, exports the competitors for each discipline.

The `BaseDiscipline.cs` file has been modified to make the `BaseDiscipline` class abstract and to include a new property `DisciplineType`. The constructor of the class has been updated to initialize this new property. A new method `ExportCompetitors()` has been added to the class, which exports the starting list of the discipline.

The `Battle.cs`, `Classic.cs`, `Jump.cs`, `Slide.cs`, and `Speed.cs` files have been updated to pass the discipline type to the base class constructor.

Finally, the `CompetitionTests.cs` file has been updated to make the `ExportResultsToCsv_ShouldCallResultsExports_ExportResultsToExcel()` test method asynchronous. The call to `competition.ExportResultsToCsv(folderPath)` has been updated to await the result.