Updated file export functionality and added folder path validation
Updated `.gitignore` to exclude `.xlsx` and `.csv` files. Added a new method `DisplayInvalidFolderPathMessage` in `ConsoleCommunicator.cs` to handle invalid folder paths. Made `Export` methods in `ExportDisciplineStartingList.cs` and `ExportRoundStartList.cs` asynchronous and added a `folderPath` parameter for file saving location. In `FreestyleSlalomCompetitionManagerRunner.cs`, added a `defaultFolderPath` field, a method to change it, and updated `ExportStartingLists` to use this path. Also updated `ExportCompetitors` in `BaseDiscipline.cs` to accept a `folderPath` parameter.