Introduced Classic discipline, removed Slide discipline
The Classic discipline was introduced to the competition manager with the addition of the `FreestyleSlalomCompetitionManager.BL.Models.Disciplines.Classic` namespace in the `FreestyleSlalomCompetitionManagerRunner.cs` file. The Slide discipline was removed from the competition manager, as indicated by the removal of the `skater.CompetitionRankSlide` line and the `Slide` discipline from the list of disciplines in the `FreestyleSlalomCompetitionManagerRunner.cs` file. The `CompetitionRankSlide` and `CompetitionResultSlide` properties were also removed from the `Competitor.cs` file, and the entire `Slide.cs` file was deleted. A new feature was added to import the starting list from an Excel file, as seen in the new `ImportStartingListXLSX.cs` file.