Updated `ImportExcelFile` return type and removed `CompetitionTests` class
Updated the `ImportExcelFile` method in the `ImportStartingListXLSX` class to return a list of `Competitor` objects instead of a `Task`. This change aligns with the method now returning the `competitors` list at the end of its execution.

Also, the `CompetitionTests` class in the `FreestyleSlalomCompetitionManager.Test` namespace has been removed. This class contained a test method for creating a `Competition` object, indicating that these tests are no longer needed or have been relocated.